update product controller

+++ src/controllers/products.ts
@@ -13,10 +13,8 @@ function handleModleError(error: any, res: Response): void {
 export const createProduct: RequestHandler = async (req: Request, res: Response, next: NextFunction) => {
     try {
         var products = await Products.create({...req.body});
-        const affectedRows = products ? 1 : 0
-        if (affectedRows === 1) {
-            return res.status(200).json({ message: "Products create successfully", data: products})
-        }
+        return res.status(200).json({ message: "Products create successfully", data: products})
+
     } catch (error) {
         return handleModleError(error, res)
     }
@@ -26,8 +24,14 @@ export const deleteProduct: RequestHandler = async (req: Request, res: Response,
     try {
         const {id} = req.params;
         const deleteProduct: Products | null = await Products.findByPk(id);
+
+        if (deleteProduct === null){
+            return res.status(404).json({message: "Products not found"})
+        }
+
         await Products.destroy({where:{id}})
         return res.status(200).json({ message: "Products deleted successfully", data:deleteProduct})
+
     } catch (error) {
         return handleModleError(error, res)
     }
@@ -36,7 +40,13 @@ export const deleteProduct: RequestHandler = async (req: Request, res: Response,
 export const getAllProduct: RequestHandler = async (req: Request, res: Response, next: NextFunction) => {
     try {
         const allProducts: Products[] = await Products.findAll()
+
+        if (allProducts.length === 0){
+            return res.status(200).json({message: "Products fetched successfully", data: "No product data"})
+        }
+
         return res.status(200).json({message: "Products fetched successfully", data: allProducts})
+
     } catch (error) {
         return handleModleError(error, res)
     }
@@ -46,7 +56,13 @@ export const getProductById: RequestHandler = async (req: Request, res: Response
     try {
         const {id} = req.params;
         const product: Products | null = await Products.findByPk(id);
+
+        if (product === null){
+            return res.status(404).json({message: "Products not found"})
+        }
+
         return res.status(200).json({message: "Products fetched successfully", data: product})
+
     } catch (error) {
         return handleModleError(error, res)
     }
@@ -57,7 +73,13 @@ export const updateProduct: RequestHandler = async (req: Request, res: Response,
         const {id} = req.params;
         await Products.update({...req.body},{where:{id}})
         const updateProduct: Products | null = await Products.findByPk(id);
+
+        if (updateProduct === null){
+            return res.status(404).json({message: "Products not found"})
+        }
+
         return res.status(200).json({message: "Products fetched successfully", data: updateProduct})
+
     } catch (error) {
         return handleModleError(error, res)
     }
